+++ docs/CHANGELOG.md
@@ -15,10 +15,16 @@ Changes are grouped as follows:
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## TBD
+### Added
+- Support for converting relationships to edges.
+
 ## [0.99.1] - 28-11-**2024**
 ### Changed
 - Remove "make_compliant" from OWL and IMF importers
 - Information rules post validation is now solely raising warnings
+- Renamed `neat.prepare.instances.relationships_as_connection` to `neat.prepare.instances.relationships_as_edges`. 
+  Changed how the conversion is done. The edge is as a node between the start and end node.
 
 ### Fixed
 - `neat.to.cdf.data_model(...)` no longer tries to deploy Cognite Models.